requirements.txt
about-time==3.1.1
alive-progress==2.4.1
altgraph==0.17.2
Appium-Python-Client==2.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
colorgram.py==1.2.0
configparser==5.2.0
crayons==0.4.0
cryptography==37.0.4
Deprecated==1.2.13
dnspython==2.2.1
dominate==2.7.0
email-validator==1.2.1
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
flask-simple-captcha==1.1.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
grapheme==0.6.0
greenlet==1.1.3
gspread==5.4.0
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
macholib==1.16
MarkupSafe==2.1.1
numpy==1.23.1
oauth2client==4.1.3
oauthlib==3.2.0
outcome==1.2.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pip-review==1.1.1
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.3.0
simplepush==1.2.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spotipy==2.20.0
SQLAlchemy==1.4.40
trio==0.21.0
trio-websocket==0.9.2
twilio==7.12.0
uritemplate==4.1.1
urllib3==1.26.11
visitor==0.1.3
webdriver-manager==3.8.2
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.1.0
WTForms==3.0.1
zipp==3.8.1
